+++ src-tauri/src/analytics/algo_controller.rs
@@ -1,16 +1,16 @@
+use super::algo_result_enums::ap::APResult;
+use super::algo_result_enums::diff_cen::DiffCenResult;
+use super::algo_result_enums::eigenvals::EigenvalsResult;
+use super::algo_result_enums::mincut::MinCutResult;
+use super::algo_result_enums::sw_cut::SWCutResult;
 use super::algo_store::AlgoStore;
+use super::algorithms::articulation_point::articulation_point;
+use super::algorithms::diffcen::diffusion_centrality;
+use super::algorithms::stoer_wagner::{recover_mincut, stoer_wagner};
 use super::algos_config::{AlgoConfig, Params};
+use super::aux_data_structures::stoer_wagner_ds::StoerWagnerGraph;
 use super::history::History;
-use crate::algorithms::articulation_point::articulation_point;
-use crate::algorithms::diffcen::diffusion_centrality;
-use crate::algorithms::stoer_wagner::{recover_mincut, stoer_wagner};
-use crate::aux_data_structures::stoer_wagner_ds::StoerWagnerGraph;
 use crate::graph::graph_ds::Graph;
-use crate::state_err_enums::ap::APResult;
-use crate::state_err_enums::diff_cen::DiffCenResult;
-use crate::state_err_enums::eigenvals::EigenvalsResult;
-use crate::state_err_enums::mincut::MinCutResult;
-use crate::state_err_enums::sw_cut::SWCutResult;
 
 pub struct AlgoController;
