+++ src-tauri/src/analytics/algo_result_enums/ap.rs
