+++ src-tauri/src/analytics/algo_result_enums/diff_cen.rs
