+++ src-tauri/src/analytics/algo_result_enums/mod.rs
