+++ src-tauri/src/analytics/algo_result_enums/most_sim_timeline.rs
