+++ src-tauri/src/analytics/algo_result_enums/pred_state.rs
