+++ src-tauri/src/analytics/algorithms/articulation_point.rs
@@ -1,7 +1,7 @@
 #![allow(dead_code)]
 #![allow(non_snake_case)]
-
-use crate::{graph::graph_ds::Graph, state_err_enums::ap::APResult};
+use super::super::algo_result_enums::ap::APResult;
+use crate::graph::graph_ds::Graph;
 use core::cmp::min;
 use defaultdict::DefaultHashMap;
