+++ src-tauri/src/analytics/aux_data_structures/cut.rs
