+++ src-tauri/src/analytics/aux_data_structures/mod.rs
