+++ src-tauri/src/analytics/aux_data_structures/neighbor_info.rs
