+++ src-tauri/src/analytics/aux_functions/convert_to_graph_from_string.rs
