+++ src-tauri/src/analytics/aux_functions/edge_factory.rs
