+++ src-tauri/src/analytics/aux_functions/mod.rs
