+++ src-tauri/src/analytics/aux_functions/take_snapshot.rs
