+++ src-tauri/src/analytics/mod.rs
@@ -1,5 +1,9 @@
 pub(crate) mod algo_controller;
+pub(crate) mod algo_result_enums;
 pub(crate) mod algo_store;
+pub(crate) mod algorithms;
 pub(crate) mod algos_config;
-pub(crate) mod global_state;
+pub(crate) mod aux_data_structures;
+pub(crate) mod aux_functions;
 pub(crate) mod history;
+pub(crate) mod state;
