+++ src-tauri/src/graph/graph_ds.rs
@@ -1,8 +1,8 @@
 #![allow(dead_code)]
 #![allow(non_snake_case)]
+use crate::analytics::algo_result_enums::eigenvals::EigenvalsResult;
 use crate::graph::edge::Edge;
 use crate::graph::node::Node;
-use crate::state_err_enums::eigenvals::EigenvalsResult;
 
 use nalgebra::DMatrix;
 use petgraph::prelude::*;
